+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import MyTrecipes from "./pages/MyTrecipe/MyTrecipes";
+import MyTrecipes from "./pages/MyTrecipes/MyTrecipes";
 import "./components/fontawesome";
 import "./App.scss";
 import ModalContainer from "./components/Modal/ModalContainer";
