+++ src/pages/MyTrecipes/TrecipeCard/trecipeCard.scss
